requirements.txt
marshmallow>=3
flask
flask-login
flask-migrate
flask-sqlalchemy
flask-static-digest
flask-marshmallow
flask-wtf
marshmallow-sqlalchemy
python-dotenv
passlib
tox
wtforms
